requirements.txt
awscli==1.19.3
boto3==1.17.3
botocore==1.20.3
click==7.1.2
colorama==0.4.3
docutils==0.15.2
fastapi==0.63.0
h11==0.12.0
jmespath==0.10.0
pkg-resources==0.0.0
pyasn1==0.4.8
pydantic==1.7.3
python-dateutil==2.8.1
python-decouple==3.4
python-multipart==0.0.5
PyYAML==5.3.1
rsa==4.5
s3transfer==0.3.4
six==1.15.0
starlette==0.13.6
urllib3==1.26.3
uvicorn==0.13.3
